fix: move formatter configs to lazy
